backend/requirements.txt
PyJWT==2.10.1
fastapi==0.124.4
langchain==1.2.0
langchain_chroma==1.1.0
langchain_community==0.4.1
langchain_core==1.2.2
langchain_experimental==0.4.1
langchain_ollama==1.0.1
langchain_text_splitters==1.1.0
langgraph==1.0.5
pandas==2.3.3
pydantic==2.12.5
pyotp==2.9.0
qrcode[pil]==8.0
uvicorn==0.38.0
